+++ backend/src/concepts/concepts.service.ts
@@ -1,8 +1,9 @@
 import { Injectable, Inject, Logger } from '@nestjs/common';
-import { Firestore } from 'firebase-admin/firestore';
-import { FIRESTORE_CONNECTION, CONCEPTS_COLLECTION } from '../firebase/firebase.module';
+import { Firestore, Timestamp } from 'firebase-admin/firestore';
+import { FIRESTORE_CONNECTION, CONCEPTS_COLLECTION, USER_CONCEPTS_SUBCOLLECTION } from '../firebase/firebase.module';
 import { GeminiService } from '../gemini/gemini.service';
-import { ConceptKnowledgeUnit } from '../types';
+import { ReviewsService } from '../reviews/reviews.service';
+import { ConceptKnowledgeUnit, UserConcept, ReviewFacet } from '../types';
 
 const CONCEPT_PROMPT = (topic: string, notes?: string) => `You are a kind and thoughtful Japanese language teacher with decades of experience. Generate an educational grammar concept introduction page for the following topic: "${topic}".
 
@@ -71,6 +72,10 @@ You MUST return a valid JSON object matching this schema exactly:
   }
 }${notes ? `\n\n**Additional notes from the teacher:**\n${notes}` : ''}`;
 
+export type UserConceptWithData = UserConcept & {
+  concept: ConceptKnowledgeUnit & { id: string };
+};
+
 @Injectable()
 export class ConceptsService {
   private readonly logger = new Logger(ConceptsService.name);
@@ -78,7 +83,124 @@ export class ConceptsService {
   constructor(
     @Inject(FIRESTORE_CONNECTION) private readonly db: Firestore,
     private readonly geminiService: GeminiService,
+    private readonly reviewsService: ReviewsService,
   ) { }
+
+  private userConceptsRef(uid: string) {
+    return this.db.collection('users').doc(uid).collection(USER_CONCEPTS_SUBCOLLECTION);
+  }
+
+  async enroll(uid: string, conceptId: string): Promise<UserConceptWithData> {
+    const existing = await this.userConceptsRef(uid)
+      .where('conceptId', '==', conceptId)
+      .limit(1)
+      .get();
+
+    const now = Timestamp.now();
+    let docId: string;
+
+    if (!existing.empty) {
+      docId = existing.docs[0].id;
+      await this.userConceptsRef(uid).doc(docId).update({ lastSeenAt: now });
+      this.logger.log(`Updated lastSeenAt for uid=${uid} conceptId=${conceptId}`);
+    } else {
+      const payload: Omit<UserConcept, 'id'> = { userId: uid, conceptId, startedAt: now };
+      const ref = await this.userConceptsRef(uid).add(payload);
+      docId = ref.id;
+      this.logger.log(`Enrolled uid=${uid} in conceptId=${conceptId} userConceptId=${docId}`);
+    }
+
+    const [ucDoc, conceptDoc] = await Promise.all([
+      this.userConceptsRef(uid).doc(docId).get(),
+      this.db.collection(CONCEPTS_COLLECTION).doc(conceptId).get(),
+    ]);
+
+    const userConcept = { id: ucDoc.id, ...ucDoc.data() } as UserConcept;
+    const concept = { id: conceptDoc.id, ...conceptDoc.data() } as ConceptKnowledgeUnit & { id: string };
+    return { ...userConcept, concept };
+  }
+
+  async findAllForUser(uid: string): Promise<UserConceptWithData[]> {
+    const snapshot = await this.userConceptsRef(uid)
+      .orderBy('startedAt', 'desc')
+      .get();
+
+    if (snapshot.empty) return [];
+
+    const userConcepts = snapshot.docs.map(d => ({ id: d.id, ...d.data() } as UserConcept));
+    const conceptRefs = userConcepts.map(uc => this.db.collection(CONCEPTS_COLLECTION).doc(uc.conceptId));
+    const conceptDocs = await this.db.getAll(...conceptRefs);
+
+    const results: UserConceptWithData[] = [];
+    for (let i = 0; i < userConcepts.length; i++) {
+      const conceptDoc = conceptDocs[i];
+      if (!conceptDoc.exists) {
+        this.logger.warn(`Concept ${userConcepts[i].conceptId} not found for user-concept ${userConcepts[i].id}`);
+        continue;
+      }
+      results.push({
+        ...userConcepts[i],
+        concept: { id: conceptDoc.id, ...conceptDoc.data() } as ConceptKnowledgeUnit & { id: string },
+      });
+    }
+    return results;
+  }
+
+  async getFacets(uid: string, conceptId: string): Promise<ReviewFacet[]> {
+    return this.reviewsService.getFacetsByKuId(uid, conceptId);
+  }
+
+  async createFacets(uid: string, conceptId: string, mechanicIndices: number[], includeAiQuestion = false): Promise<{ created: number }> {
+    const conceptDoc = await this.db.collection(CONCEPTS_COLLECTION).doc(conceptId).get();
+    if (!conceptDoc.exists) throw new Error(`Concept ${conceptId} not found`);
+
+    const concept = { id: conceptDoc.id, ...conceptDoc.data() } as ConceptKnowledgeUnit & { id: string };
+
+    const facets: Array<{ kuId: string; facetType: ReviewFacet['facetType']; data: any }> = [];
+
+    for (const idx of mechanicIndices) {
+      const mechanic = concept.data.mechanics[idx];
+      if (!mechanic) {
+        this.logger.warn(`Mechanic index ${idx} out of range for concept ${conceptId}`);
+        continue;
+      }
+      facets.push({
+        kuId: conceptId,
+        facetType: 'sentence-assembly',
+        data: {
+          mechanicIndex: idx,
+          goalTitle: mechanic.goalTitle,
+          fragments: mechanic.naturalExample.fragments,
+          answer: mechanic.naturalExample.japanese,
+          english: mechanic.naturalExample.english,
+          accepted_alternatives: mechanic.naturalExample.accepted_alternatives ?? [],
+          sourceId: conceptId,
+          sourceTitle: concept.data.title,
+        },
+      });
+    }
+
+    if (includeAiQuestion) {
+      facets.push({
+        kuId: conceptId,
+        facetType: 'AI-Generated-Question',
+        data: { topic: concept.data.title, sourceId: conceptId, sourceTitle: concept.data.title },
+      });
+    }
+
+    const created = await this.reviewsService.createFacetBatch(uid, facets);
+
+    await this.userConceptsRef(uid)
+      .where('conceptId', '==', conceptId)
+      .limit(1)
+      .get()
+      .then(snap => {
+        if (!snap.empty) snap.docs[0].ref.update({ facetCount: created });
+      });
+
+    this.logger.log(`Created ${created} facets for uid=${uid} conceptId=${conceptId}`);
+    return { created };
+  }
 
   async generate(uid: string, topic: string, notes?: string): Promise<ConceptKnowledgeUnit & { id: string }> {
     this.logger.log(`Generating concept for topic="${topic}" uid=${uid}`);
